Improve database migration support
This implements a handful of fixes to better support migration of old databases
(pre-version 3). The main things were a couple of new CLI entry points for
extracting and injecting inputs from/into a database file. This is to support
file-based migration, so that when loading the database, a reactor model can be
fashioned from the inputs stored in the database. We may also want to work on
the input loading to support initialization and migrations purely through
streams, avoiding the need to extract files for migration. This will require
more rework. We also add some better resiliency to strange database states that
would otherwise crash during conversion, sped up parameter renames, and taught
the reactor factory to honor the optional geom= kwarg. Going forward, we may
also wish to cast the migration of the geom stuff by converting a passed
`SystemLayoutInput` to a valid grid blueprint, rather than creating a reference
to the geomFile in a stub grid blueprint.