Replace object references with soft links
This replaces the use of object references to handle HDF5 attributes
that are too large to store natrually. This is because object references
break when copying an object that contains them to a new HDF5 file, as
is done in mergeHistory(). Instead, we use a simple string to refer to
the targed dataset like "@[path/to/dataset]".