Ensure that Cases do not consider themselves to be their own dependency
This makes sure that self is not contained in a Case's set of
dependencies. This can happen sometimes when a plugin is infering a Case
dependency based on a referenced input file name. If that name maps back
to the current case, it would errantly flag the current Case as its own
dependency.