Add BOL HM mass  parameter. (#24)
* Remove superfluous passing of bolBlock to completeInitialLoading()

Callers were passing self in as the bolBlock. self is used by default
anyways.

* Resolve formatting issues

* Add BOL HM mass parameter

This is a bit wasteful, but for now is the easiest way to reliably get
HM mass (or anything for that matter) at BOL. Future work is to find a
way to handle these in some way that doesn't require a whole parameter.