Uniform mesh conversion fixes (#51)
* Improve uniform mesh conversion when doubled up.

The uniform mesh converter would take the source reactor and create a
new reactor with a uniform mesh and simply update the number density
parameters on it. After the uniform calculation was done (often a global
flux calculation updating power and flux), a bunch of neutronics
parameters were mapped back to the original reactor. This worked fine
when there was just one conversion like this happening in a sequence.

However, if two similar conversions happen in a run (e.g. when there's a
neutron-only flux calculation followed by a gamma-transport fixed-source
problem), then the gamma problem would have invalid power parameters on
the converted reactor.

This change upgrades the uniform mesh converter to map all relevant
parameters from the source reactor to the converted reactor in the first
place, making it symmetric with the mapping from the converted reactor
back that has always happened. Now, power, flux, and related parameters
are correct on a 2nd or later pass of this converter at the same
timestep.

* Made neutronics-specific uniform mesh subclass.

* Moved a bunch of block parameter definitions into neutronics.