Rearrange how tutorial notebooks are handled
The jupyter notebooks used in the tutorials depended on a handful of
input files, which are located in the ARMI source tree. Relative paths
were being used to refer to these input files, meaning that the
notebooks could not be used from any other location, which is needed to
resolve race conditions when running tests. This adds some code to the
doc/conf.py to simply copy them into the appropriate location so they
can be used from the working directory.