Add location-based history tracking on Database3
This adds the `getHistoryByLocation()` and `getHistoriesByLocation()`
methods to the Database3 class. These are quite similar to their
non-location based siblings, but rather than following an ArmiObject in
the database by serialNum, it does a little extra work to figure out
each object's location, then pull data from the object that is actually
at that location for each time node requested. Other changes:
 - Added some type annotation stuff
 - Added `computeParents()` classmethod to Layout to help make sense of
 the hierarchy without having to fully reconstruct.
 - Added a `getAncestorWithDistance()` method to ArmiObject. This is
 useful for determining not only the ancestor object, but how many
 layers above it happens to be.
 - Moved some methods that should obviously be Composite methods out of
 ArmiObject. There are plenty more that are miscategorized and still
 need moved, but that should be its own thing.