Fixes to Circular Ring HEX to RZ Conversion and Preloading ISOTXS on Core (#79)
* A bug was found using the circular ring mode when converting between
the hex reactor to the rz reactor where total, fissile, and heavy metal
masses were not always being preserved. The circular ring mode conversions
were never used in practice, but having this available as an option could
lead to future issues for users.

Additionally, the `hexRingGeometryConversion` setting is being removed
since this is implied as the default behavior in `getAssembliesInRing`
on the core. This default is overwritten using the `circularRingMode`
setting.

Finally, the core.lib was assuming the existence of an `ISOTXS` file
in the working directory if it was not explicitly assigned. This is
generally not good practice as a cross section set from previous core
state could be accidently read in without the user realizing it. This is
being changed to happen explicitly only when `genXS` is not set versus
always implicitly being loaded. The library loading still occurs when `core.lib`
is called.