Add tutorial on making ARMI applications
We had some documentation and examples of making ARMI-based apps, but
needed a real walkthrough tutorial. Using the famous Django tutorial as
a guide, we built one that makes an entire app with a plugin and a few
physics solvers (albeit fake ones). The tutorial includes things most
users will need, like adding custom materials. We will extend it with
more settings and parameters later.

Along the way there were a few issues that popped up for resolution,
including:

* There was a file-name case sensitivity issue discovered while building
  documentation in linux.

* An erroneous detail assembly could be added in ring 1, position N
  without error. Now it is an error when N>1

* Building documentation several times led to a "file too long" error
  because of recursion in the `_build` folder. This was fixed by adding
  an exclusion.