Add run-suite entry point for local param sweeps
A run-suite entry point is now provided, similar to the submit-suite
entry point available in some apps. This entry point runs all cases of a
suite one after the other, on the local machine. It is useful for
situations where a HPC is not available but a big workstation is.

Relatedly, some issues with paths in the parameter sweep system were
dealt with by allowing plugins to provide relative glob patterns that
are expanded at a higher level. This simplifies some plugin code (e.g.
when lattice physics specifies ISOTXS files) and makes the param sweep
system more capable.