Improve merging behavior of GAMISO and PMATRX files (#98)
When handling GAMISO and PMATRIX files for merging into ISOTXS, and when
a custom suffix is provided, library files without a suffix were being
ignored, even if a library with the supplied suffix did not exists. This
changes the logic to fall back to a file without a suffix, which should
properly handle situations where the suffix is not relevant (e.g., when
voided cross sections aren't needed for gamma transport or whatever).