Add special flag to allow double-configure
This flag is changed by Sphinx during documentation build.
The reason for it is to allow examples in the gallery to be completely
copy/pasteable by users while also being executed during documentation
builds after the initial configuration. Another option used elsewhere is
to use `if armi.isConfigured()` guards, but for early example we prefer
to not confuse users with that kind of detail.