Stabilize "refAssem" (#118)
This explicitly defines the Core's "refAssem" to be the center-most FUEL
assembly, if there are fueled assemblies, or the center-most assembly if
there are not. It also bends some code that was doing essentially the
same thing towards using `core.refAssem` so that their behavior is at
least consistent.

In support of this, an optional `sortKey=` argument was added to the
`getAssemblies()` method to inject a sorting key. This allows client
code to sort assemblies however they wish, without having to resort
re-sorting the already-sorted Assemblies.