Define and implement plugin hook to handle MPI resets
For ARMI runs in parallel over MPI, several internal plugins were
imported, causing ImportErrors. The rectify this, we define here
a new plugin hook, mpiActionRequiresReset, and implement it
as appropriate on the framework and in other plugins.

Along the way, it was noticed that if a worker MPI node fails
in this way, the traceback was not available. So we added an
additional traceback printout to help aid in debugging. Still
not quite right is the fact that the run hangs in this situation.

Sidenote: No one is too happy with the MPI resetting operation
in general. It is a remnant of challenging memory issues related
to gigantic long-running cases.