Remove old block homogenized ndens params (#128)
While number density state is stored on components, a cache of the
block-homogenized number densities was maintained in an ad-hoc fashion
in a long series of block parameters. The only known use case for these
is to quickly check spatial distributions of composition in
visualization tools that cannot show state below the block level.

This change removes the block params from the reactor model. However,
it does still add them to the HDF5 database by computing them on the fly
during DB writing. This is intended to be temporary and should be
released once suitable alternative visualization techniques are
available.