Remove the geom.xml file from the unit test rx
The armiRun.yaml inputs are the basis of many unit tests that need a
reactor. They have been using the mostly-deprecated geom.xml format for
the core map. This change converts that to grid asciimap and loads it
through the blueprints !include mechanism.

Along the way, some PARTISN-specific inputs and params were removed.