Add support for converting old database flags
Before this, if flags are added to an application, databases from a
previous version would be unreadable, because the Flags serializer makes
sure that the meaning of all flags is the same between the reading app,
and the app that wrote the database. Now, it will only check that all
old flags still exist, but permit converting to the new Flag numeric
values.