Remove None option in xs settings to avoid new err
The schema checking dependency added a sorting function that causes one
of our schemas to fail because a None can't be sorted among strings.
Rather than fixing the dependency, I just took the None option out
because its justification in that list is not clear or documented, and
also the tests pass without it.

Also added a simple unit test that ensures the main armi entry point
executes with no arguments as a result of a recent issue preventing it
from doing so.