Change geomType arguments to prefer GeomType enum (#229)
Most code that takes a geomType argument used to expect a string, which would
be tested against the string constants defined in the geometry module. This
was error-prone, inefficient, and ambiguous. This adds some extra conversion
functionality to the GeomType enumeration class and changes most interfaces that
pass geomType around to use instances of the GeomType enum instead. Some
functions will still attempt a conversion from string to aid in backwards
compatibility, but these should be reduced moving forward.

Additionally, to better-organize the code in the geometry module, the
SystemLayoutInput class has been migrated to its own module.