Add the first GUI test for the grid editor.
This commit adds the `GuiTestCase` class, as well as an example UI test.
At the start of each test, a new instance of the `GridBlueprintControl` class
is created and displayed, and the test author can define the sequence of
actions to be simulated using `wx.UIActionSimulator`, and check that the
properties of the `GridBlueprintControl` instance are as expected. The test
can also be run in headless mode with `pytest-xvfb`.

Unfortunately, this test only works in certain environments (we've only
gotten it to work locally on Linux with pure X11 environments), so we
are disabling the tests for now.