Enable copying of interface inputs, even if interface is disabled
There is no clear reason to skip copying interface files when the
interface in question is disabled, other than to perhaps save space on
large files. This was specifically leading to missed fuel handler code,
since the shuffleLogic file needs to be present regardless of whether it
is going to be enabled.