Improve interface input copying/setting updating (#241)
This improves the handling of interface files when cloning Cases and
Case Suites. The most important change is support for adjusting cloned
settings to point to the target file.

This also removes the logic to skip copying interface files when the
interface in question is disabled. There is no clear rationale to this,
other than to perhaps save space on large files. This was
specifically leading to missed fuel handler code, since the
shuffleLogic file needs to be present regardless of whether it is 
going to be enabled.