Eliminate axial and ring limitations of names and locations (#239)
The blocks used to have an axial character indicating axial height,
as well as "ring number" for both locations and names, 
but this is inherently limited. That behavior is now delegated
to the plugins that need it so that more general inputs can be created.

Fixes #233.

* Update code and tests to work with new location labels
* Added a migration that can update location-based settings
* Add new unit test for migration

* Fix false positive possibility in test util

The utility used to compare line-by-line was iterating over the expected
file. However, in situations where the first x lines are identical but
then there are more lines, this test would give a false positive. Now
it will fail if either the test or expected file have additional lines.