Bump h5py version
h5py changed the manner in which they handle strings, requireing some
modifications to the database3 implementation. These were relatively
straightforward, but the new h5py required a newer version of numpy,
which unerthed some new warnings and issues in the tests. This addresses
all of those as well.

Of particular note, is the addition of floating-point epsilon-aware
line-by-line file comparisons. This is needed, since newer numpy
versions produce slightly different results in certain scenarios, and we
want to make the tests a little less seensitive to minor differences
between numpy versions. The approach is to (in the event that string
comparison of two lines fails) split each line into words along
whitespace. For each pair of words, attempt parsing them as floats. If
they can both be parsed into floats, do a relative epsilon comparison.
Any trouble in this process and it will declare the strings non-equal.