Broaden setting parsing error checking
Loading a settings file that had a setting with an invalid value was not
getting caught here because it was just a single invalid. This broadens
the exception catching to capture any related validation error.

Additionally the directory changer was changed to not make a separate
error folder since most of the time migration type errors will be found
in a local setting rather than while running on HPC.