Defer fast path resolution
Though we moved fast path resolution to a function to prevent
import-time execution of it, this one case still ran at import time
because it was in a class definition. As a result, any client of the
TemporaryDirectoryChanger would be using the CWD as the fast path rather
than some probably-better local fast path. In cases on clusters where
cwd is a shared network drive, this could really lower performance.