Make master cs consistent in gallery example
The global master case settings object was causing an error during
documentation building in the code that checks that all cs objects in
the tree are consistent. This fixes the issue. Note however that we do
hope to eliminate the concept of a global CS and at that point will
eliminate the check that failed, and revert this commit.