Fix incorrect mult for non-grid block components (#255)
All components in blueprints-defined blocks that had grids assigned but
were not themselves in the grid got a multiplicity of zero, which made
the volume fractions (and therefore everything else) wrong. This fix
detects when a component is not represented in the grid and just leaves
the user-input multiplicity. A unit test covering the situation was also
added.

Fixes #254.