Clean up plotFlux.
plotFlux needs to be cleaned up, and eventually moved to
utils.plotting. This commit abstracts some of the repetitive functions
that existed in plotFlux into a new class, BlockListFlux. BlockListFlux
exists only within the scope of plotFlux. This abstraction reduces
the size of the main function and makes it easier to read.

Also add tests for "peak" and "bList2" cases. Can't test adjoint at the
moment because the testBlock doesn't have an adjoint flux.

The function is still relatively long because it handles several
different cases: whether the second set of data should be (a) adjoint
flux or (b) flux from a different list of blocks. In any case, there is
also an option whether or not to plot the peak flux. It would be
difficult to simplify or shrink this function any further without breaking
out the separate cases into separate functions, which wouldn't be all that
helpful or clarifying.