Move plotFlux from Block static method to utils.plotting.
plotFlux is a better fit in utils.plotting. plotFlux remains as a
pass-through function in the Block namespace that passes the arguments
along to the actual implementation in utils.plotting. Keeping this
function available here will minimize the side effects from removing a
function that plugins might be expecting here. Once all pluging calls to
this function have been updated to directly access the actual
implementation, the pass-through function can be safely eliminated. This
pass-through function is effectively a symbolic link.