Add unit tests for materials.
Needed more test coverage for material.py. Identified one issue, with
the implementation of isBeyondIncubationDose and
updateDeltaDPApastIncubation. The logic tries to evaluate the result
of self.modelConst["Rincu"]. If "Rincu" is not in modelConst, it will
not return False, but rather throw a KeyError. First, we need to check
whether it is in the dict, then we can proceed.