Add unit test for setPinPowers and fix bug. (#260)
setPinPowers has a boolean option gamma. If gamma == True, then it adds
the gamma pin power to the neutron pin power to get the total pin power,
and stores in in b.p.pinPowers. The old code used the following
notation:

self.p.pinPowers = self.p.pinPowersNeutron + self.p.pinPowersGamma

This would work for numpy arrays, but the pin powers are defined as
python lists in this method. As a result, python was concatenating a
list of the gamma powers onto the neutron pin power list instead of
computing an element-wise addition.