Clean up Block.plotFlux and move it to utils.plotting. (#261)
* Clean up plotFlux.

plotFlux needs to be cleaned up, and eventually moved to
utils.plotting. This commit abstracts some of the repetitive functions
that existed in plotFlux into a new class, BlockListFlux. BlockListFlux
exists only within the scope of plotFlux. This abstraction reduces
the size of the main function and makes it easier to read.

Also add tests for "peak" and "bList2" cases. Can't test adjoint at the
moment because the testBlock doesn't have an adjoint flux.

The function is still relatively long because it handles several
different cases: whether the second set of data should be (a) adjoint
flux or (b) flux from a different list of blocks. In any case, there is
also an option whether or not to plot the peak flux. It would be
difficult to simplify or shrink this function any further without breaking
out the separate cases into separate functions, which wouldn't be all that
helpful or clarifying.

* Move plotFlux from Block static method to utils.plotting.

plotFlux is a better fit in utils.plotting. plotFlux remains as a
pass-through function in the Block namespace that passes the arguments
along to the actual implementation in utils.plotting. Keeping this
function available here will minimize the side effects from removing a
function that plugins might be expecting here. Once all pluging calls to
this function have been updated to directly access the actual
implementation, the pass-through function can be safely eliminated. This
pass-through function is effectively a symbolic link.

* Move test for Block.plotFlux over to test_plotting.

Block.plotFlux was moved to utils.plotting.plotBlockFlux. This commit
moves the associated test over to test_plotting.