Avoid hard-coded unit test references in test_materials.
Some of the functions in materials required hard-coded reference results
that depend upon the physical properties of the materials on which
they're called. If these properties change, the tests would also need to
change. Since the functions are relatively short, we replaced the
hard-coded value with a dynamic calculation of the correct value. In
other words, the function is re-implemented in the test.