Force pinPowers parameters to be numpy arrays for faster operations. (#265)
setPinPowers() in blocks.py was setting pinPowers as python lists.
Floating point operations on these lists, such as the pin-wise addition
of gamma and neutron powers, are much faster with numpy arrays than
python lists, and the semantics of various arithmetic operators differ
between the two. Force these parameters to numpy arrays with a setter.