Avoid plotting empty vectors in keffVsTime
Sometimes a database will be populated with an empty vector for the
uncontrolled multiplication factor. I believe this happens if the
value of keffUnc is not updated during an ARMI run. This leads to a
list of zeros being passed to the bookkeeping keffVsTime graph which
_technically_ evaluates to `True` and gets plotted. Using `any` for
the conditional check skips plotting empty lists and `None`s.

Signed-off-by: Andrew Johnson <drewejohnson@users.noreply.github.com>