Add/improve unit tests for Material. (#266)
* Add unit tests for materials.

Needed more test coverage for material.py. Identified one issue, with
the implementation of isBeyondIncubationDose and
updateDeltaDPApastIncubation. The logic tries to evaluate the result
of self.modelConst["Rincu"]. If "Rincu" is not in modelConst, it will
not return False, but rather throw a KeyError. First, we need to check
whether it is in the dict, then we can proceed.

* Avoid hard-coded unit test references in test_materials.

Some of the functions in materials required hard-coded reference results
that depend upon the physical properties of the materials on which
they're called. If these properties change, the tests would also need to
change. Since the functions are relatively short, we replaced the
hard-coded value with a dynamic calculation of the correct value. In
other words, the function is re-implemented in the test.

* Return False if no incubation dose is present.

If the material does not have a "Rincu", return False from
isBeyondIncubationDose.

* Simplify logic in isBeyondIncubationDose.

* Simplify logic in updateDeltaDPApastIncubationDose.

Removed a redundant check to isBeyondIncubationDose.
updateDeltaDPApastIncubationDose is only called after
isBeyondIncubationDose evaluates as true, so the check does not need to
be repeated within updateDeltaDPApastIncubationDose.