Fix printDensities and add a unit test.
printDensities was attempting to format a nuclide object into a
string, instead of the nuclide's name attribute. This was caught
when I tried to create a unit test for the function.

Wanted to test the log output created by printDensities, but it
doesn't seem like there is a good way to do it. Python has a
context manager for capturing output from native logging.Logger
types, but it seems like there's no parallel capability that would
work with our runLog class.