Fix KeyError in asciiMapFromGeomAndSym (#273)
Use the geometry.GeomType.fromAny() method to input a string or
enumeration representation of a geomType and return the enumeration.

Symmetry does not have enumeration at the moment, so we'll continue to
use the string representation for now. Defining a single enumeration
for symmetry would be complicated by the different types of options
(full/third/quarter, reflective/periodic, etc.) that can appear in
combination with one another.