Introduce enumerations for symmetry.
GeomType had been previously converted to use an enumeration, while
still maintaining the string representation for interfacing with yaml
and databases. Symmetry options, of which there are many, were still
stored in a string and interpreted with string operations.

This commit introduces a new class in geometry.py, SymmetryType. This
class handles three types of symmetry options:
1. "shape": full, third, quarter, eighth, or sixteenth core
2. "boundary": reflective, periodic, or no symmetry
3. "isThroughCenter": boolean to indicate whether the symmetry line
   passes through the center assembly

Enumerations were introduced for shape (ShapeType) and boundary
(BoundaryType), while isThroughCenter is a simple boolean option.
The enumerations are configured in the mold of pre-existing the
GeomType enumeration. The enumerations can be constructed from a
string representation, but they also have an agnostic constrcutor
fromAny() that can accept either the string or the enumeration, and
return the enumeration. The SymmetryType class is effectively a wrapper
that combines the enumerations in a defined way.

fromAny() is helpful for avoiding type clashes and needing to
meticulously track where conversion between string and enumeration is
happening. It also facilitates storing the symmetry as a string-type
property of the Grid class, where the setter ensures that it can only
be set to a valid state.

No significant optimization effort has been exerted in this commit.
In theory, the code introduced here could be faster than the string
parsing that it replaces, but since the enumerations still frequently
deal in strings we should probably be careful. The SymmetryType class
is not quite as lightweight as the plain string representation it
replaces, but it should be less error-prone and hopefully easier to
maintain. If performance degradation is an issue, we may need revisit
and be more careful about where and how SymmetryType objects are
constructed/used.