Check if SymmetryType is valid before returning.
A user could theoretically generate any combination of shape, boundary,
and isThroughCenterAssembly by passing a string or enumeration to the
constructors of SymmetryType. This commit adds a returnIfValid() method
that only returns the constructed object if it matches one of the valid
options defined in VALID_SYMMETRY. If not, raise a ValueError.