Use try-catch to avoid invalid symmetry error.
In HexBlock.getSymmetryFactor(), previous approach was to independently
check that self.parent exists and self.parent.spatialLocator.grid exists,
and then try to make a SymmetryType object from
self.parent.spatialLocator.grid.symmetry. If any of these things fails,
we just want to return 1.0, so we don't really need complicated logic.
Construct the SymmetryType inside a try statement, and return 1.0 if the
try fails. If it doesn't fail, we can execute the rest of the code.