Fix an error introduced in gridBluePrint.py.
I introduced a bug while trying to simplify logic in gridBluePrint.py.
I removed a check for the existence of self.symmetry because it is set
to a default value in the base constructor, so I assumed it would always
exist. I didn't notice that there is another constructor that doesn't
set the symmetry attribute, so it is often None.