Use the SymmetryType constructor to build strings.
Symmetry strings are used for comparison all over the code and test
base. Instead of building the strings inline, they are built by passing
the enumerations to the SymmetryType constructor and returning the
official string representation. This keeps the string-building format
and logic in a single location.

Fix a couple of bugs for NO_SYMMETRY case.

There was some confusion about whether to return the string "no
symmetry" or an empty string "" when there is no boundary symmetry
condition.