Use tuple representation to check validity of SymmetryType.
Tuple of enums is a better representation of a SymmetryType than a
string. An __iter__ method was defined for SymmetryType so that a
tuple representation can be used to check against the list of tuples in
VALID_SYMMETRY. validSymmetryStrings() method was defined to convert the
tuples to a list of strings, which makes error messages more helpful for
users.

With tuple option for SymmetryType instantiation, we can move away from
constructing with fromStr() in favor of constructing with __init__()
using the enums directly, or with tuples of the enums passed to
fromAny().

Move checkValidGeomSymmetryCombo() from geometry.py to
settingsValidation.py, where it is used.