Refactor to use SymmetryType default constructor in classmethods.
Fold the _returnIfValid() function into __init__(); refactor the
classmethods that return a SymmetryType object to call the default
constructor so that the validity check is always performed.

Checking for validity directly in __init__() means we can remove the
validity check query in settingsValidation.py. The validity of the
symmetryType is still checked indirectly, when it is constructed during
the call to checkValidGeomSymmetryCombo(). checkValidSymetry() can be
inlined in __init__() and removed.