Implement __eq__ and __hash__ for SymmetryType.
Direct comparison of SymmetryType that were ostensibly equivalent was
producing a False. The __eq__ method was implemented to directly check
equivalence of the tuple representations of two SymmetryType. __hash__
is defined to hash the tuple representation, which will ignore any
superfluous differences between two SymmetryType instances that are not
relevant to the symmetry options.