Add enumeration for symmetry types. (#276)
GeomType had been previously converted to use an enumeration, while
still maintaining the string representation for interfacing with yaml
and databases. Symmetry options, of which there are many, were still
stored in a string and interpreted with string operations.

This commit introduces a new class in geometry.py, SymmetryType. This
class handles three types of symmetry options:
1. "domain": full, third, quarter, eighth, or sixteenth core
2. "boundary": reflective, periodic, or no symmetry
3. "isThroughCenter": boolean to indicate whether the symmetry line
   passes through the center assembly