Add tests for lumpedFissionProducts. (#272)
* Add unit test for getNumberFracs in lumpedFissionProduct.

Add a unit test to run the getNumberFracs method of
lumpedFissionProduct. The docstring for getNumberFracs,
and for getMassFrac, suggested that returned dict is
indexed by nuclide name, but it is actually indexed by
a nuclide object.

* Fix printDensities and add a unit test.

printDensities was attempting to format a nuclide object into a
string, instead of the nuclide's name attribute. This was caught
when I tried to create a unit test for the function.

* Add unit test for getGasRemovedFrac.

* Add unit test for LumpedFissionProductCollections.

Add unit tests for getNumberDensities and getMassFrac methods of
LumpedFissionProductCollection. Fixed one bug in
lumpedFissionProductCollection's getMassFrac.

* Simplify the printDensities function for LumpedFissionProducts.

* Specify black version for GitHub workflow.

The formatting check on GitHub was failing to find a stable version of
black.