Blueprint migration bug fix. (#280)
Use GeomType enumeration instead of string in geom auto-migration

GeomType bug in #278 was not caught by unit tests because
test_rz_reactors.py was passing despite the bug. Modify ThRZGeom.yaml to
make test_rz_reactors.py so that it only passes when the if block in
question (in the blueprints/__init__.py migrate() function) is correctly
executed.