Fix encoding of images in HTML reports (#282)
After an ARMI run, none of the images in the `reports/*html` are
valid. Looking at the leading string in some of the reports, I see
`data:image/jpg;base64,b'/9j/4...`. When I pipe the output of the
image through `base64` and paste that into a browser after
`data:image/jpg;base64`, I see `data:image/jpg;base64,/9j/4A...` and
the image is displayed.

This decodes the binary string after `base64` encoding to remove the
leading `b'` in the string representation.

Signed-off-by: Andrew Johnson <a.johnson@usnc-tech.com>