Unit test to catch modulus error during over-rotation.
If the sum of the current rotation state and the additional rotation
exceeds 6, the old rotatePins function would over-rotate. The unit
test was modified to induce this situation, where a 4 x 60 degree
rotation is applied to a block that has already been rotated 4 x 60. In
this case, the old rotatePins would calculate a rotNum of 8, which would
make some pin indexes incorrect.