py2to3 - removing future print import
Several files still had this line at the top:
from __future__ import print_function
Generally, this code is only useful in Python v2.

There are only a tiny amount of edge cases where
this line is useful in Python v3, mostly having
to do with trying to connect to Python v2 modules.
But it doesn't appear to me this is the case.