The `_RZThetaReactorMeshConverterByRingComposition` class was incorrectly using the `core.getAssembliesInRing` to obtain the assemblies within a given ring. This class is intended to be used for non-circular geometries, like HEX or Cartesian, so the method call is now set to `core.getAssembliesInSquareOrHexRing`.
This change is made to decouple the behavior of this class
with the `circularRingMode` setting. Otherwise, the behavior
of the `HexToRZThetaConverter` class changes depending on whether
or not the user sets `circularRingMode` to True or False. The
`circularRingMode` is intended for use in fuel management and
not in this conversion process.