Complete removal of locations.py (#264)
This removes the existing `locations.py` module in the framework since
it was planned to be replaced with the newer implementation of grids.
There were only a few instances in the framework that still relied on 
the older locations API, mainly in `blockConverters.py`. The documentation
in the block converters was changed to include some more information 
about the assumptions that it must make about the grids to function.

Additionally, this change also removed any specialization that `HexAssembly`
and `CartesianAssembly` have over a normal `Assembly` class. Ideally, the
base assembly would implement the required functionalities, but the `RZAssembly`
and `ThRZAssembly` classes still implement special functions to handle some
mesh-related attributes/parameters.